🚚 Default to most simple NLB direct connection.
This modifies the existing default configuration to not prefer
additional VPC Endpoints in the API Account. However, you can just add
`vpc_endpoint_local = true` and have the same behaviour. This is still
required as a first step to use a connectivity account to trigger the
generation of the VPC Endpoint Service toward the NLB.

In addition, this is much better documented, and includes sequence
diagrams that should make the call chains a bit more clear.